fix: make every update recipe field optional

+++ src/recipes/recipes.service.ts
@@ -77,7 +77,10 @@ export class RecipesService {
       throw new NotFoundException();
     }
 
-    if (recipe.imageKey !== updateRecipeDto.imageKey) {
+    if (
+      updateRecipeDto.imageKey &&
+      recipe.imageKey !== updateRecipeDto.imageKey
+    ) {
       await this.uploadFileService.deleteImage(recipe.imageKey);
     }
